+++ app/src/main/java/io/itch/fr/quizgame/feature_start/domain/GetQuestionsUseCase.kt
@@ -2,8 +2,9 @@ package io.itch.fr.quizgame.feature_start.domain
 
 import io.itch.fr.quizgame.feature_quiz.data.entities.Question
 import io.itch.fr.quizgame.feature_quiz.data.repositories.QuestionRepository
+import javax.inject.Inject
 
-class GetQuestionsUseCase(private val questionRepository: QuestionRepository) {
+class GetQuestionsUseCase @Inject constructor(private val questionRepository: QuestionRepository) {
     suspend operator fun invoke(): List<Question> {
         return questionRepository.getQuestions()
     }